Bump up nvComp version to 5.3.0.16 (#6430)
* Fixes test_dali_variable_batch_size.test_inflate, which compressed int64
  samples while relying on the Inflate default output type of uint8.
* Adds the optional check_output_size argument to decoders.inflate.
* When enabled, it validates the decoded LZ4 size before decompression and
  raises an error if the requested output buffer is too small.

+++ dali/test/python/test_dali_variable_batch_size.py
@@ -1460,7 +1460,7 @@ def test_inflate():
         def piepline():
             defalted = fn.external_source(source=input_data)
             shape = fn.external_source(source=input_shape)
-            return fn.decoders.inflate(defalted.gpu(), shape=shape)
+            return fn.decoders.inflate(defalted.gpu(), shape=shape, dtype=types.INT64)
 
         return piepline(batch_size=max_batch_size, num_threads=4, device_id=0)
